Add 101_add_missing_link_libraries.patch
Loading imgsvg was causing segfaults and saying it was missing a symbol
from libdecoration. I checked the buildlogs and found:
.
libimgsvg.so contains an unresolvable reference to symbol decor_apply_gravity: it's probably a plugin.
libopengl.so contains an unresolvable reference to symbol dlsym: it's probably a plugin.
libcompiztoolbox.so contains an unresolvable reference to symbol XRenderFindVisualFormat: it's probably a plugin.
.
These 3 files are obviously plugins and dpkg-shlibdeps assumes that the program
that loads the plugins provides theses symbols therefore not giving an error.