Reenable the tests.
Upstream fixed gtest but google mock still needs to be compiled in Debian.
This will probably get fixed when google mock in Ubuntu stops shipping their
libraries like gtest did.
.
The cmake script already disabled the test if gtest/gmock was not found.
Remove the override so that when they fix this upstream the test work.